added some new features

+++ code/que/daemon.py
@@ -1,22 +1,14 @@
-from typing import Optional, Union, TypeGuard, Dict, Any, TypedDict
-from pathlib import Path
-from multiprocessing import Process, Event
+from typing import Optional
+from multiprocessing import Process
 from multiprocessing.synchronize import Event as EventClass
 from logging import Logger
-import json
 import os
 import time
 
 from .core import (
     connect_manager,
-    DaemonState,
     DaemonStateHandler,
-    WR_LOG_PATH,
-    WR_PATH,
-    WR_MODULE_PATH,
-    DN_LOG_PATH,
-    DN_NAME,
-    DAEMON_STATE_PATH,
+
 )
 from .worker import Worker
 
@@ -28,12 +20,14 @@ class Daemon:
         worker: Worker,
         logger: Logger,
         local_state: DaemonStateHandler,
-        stop_event: EventClass,
+        stop_worker_event: EventClass,
+        stop_daemon_event: EventClass,
     ) -> None:
         self.worker = worker
         self.logger = logger
         self.local_state = local_state
-        self.stop_event = stop_event
+        self.stop_worker_event = stop_worker_event
+        self.stop_daemon_event = stop_daemon_event
         self.worker_process: Optional[Process] = None
         self.supervisor_process: Optional[Process] = None
         
@@ -63,22 +57,25 @@ class Daemon:
                 return False
             
             # Small backoff before restarting to prevent rapid looping on hard crashes
-            if not self.stop_event.is_set():
+            if not self.stop_daemon_event.is_set():
                 self.logger.info("Restarting worker in 1 second...")
                 time.sleep(1.0)
+            else:
+                self.logger.info("Stop event detected, not restarting worker.")
+                return False
                 
         return True
 
-    def hard_cleanup(self) -> None:
+    def hard_cleanup(self, supervisor: bool = True, worker: bool = True) -> None:
         """
         Forcefully terminate the worker and supervisor processes if they are running.
         """
-        if self.worker_process and self.worker_process.is_alive():
+        if worker and self.worker_process and self.worker_process.is_alive():
             self.logger.info("Forcefully terminating worker process...")
             self.worker_process.terminate()
             self.worker_process.join()
 
-        if self.supervisor_process and self.supervisor_process.is_alive():
+        if supervisor and self.supervisor_process and self.supervisor_process.is_alive():
             self.logger.info("Forcefully terminating supervisor process...")
             self.supervisor_process.terminate()
             self.supervisor_process.join()
@@ -99,9 +96,9 @@ class Daemon:
 
         self.logger.info(f"Supervisor loop started. PID: {os.getpid()}")
 
-        while not self.stop_event.is_set():
+        while not self.stop_daemon_event.is_set():
             try:
-                self.worker_process = Process(target=self.worker.start, args=(self.stop_event,))
+                self.worker_process = Process(target=self.worker.start, args=(self.stop_worker_event,))
                 self.worker_process.start()
                 
                 state_proxy.set_worker_pid(self.worker_process.pid)
@@ -114,7 +111,7 @@ class Daemon:
 
             except Exception as e:
                 self.logger.error(f"Supervisor error: {e}")
-                if self.stop_event.is_set():
+                if self.stop_daemon_event.is_set():
                     break
                 time.sleep(1.0) # Prevent tight loop on error
 
@@ -131,7 +128,8 @@ class Daemon:
             self.logger.warning("Supervisor is already running.")
             return
 
-        self.stop_event.clear() # Reset event in case it was set previously
+        self.stop_daemon_event.clear() # Reset event in case it was set previously
+        self.stop_worker_event.clear()
         self.local_state.set_awake(True)#child inherits local state
         self.local_state.to_disk() 
 
@@ -140,14 +138,37 @@ class Daemon:
          
         self.logger.info(f"Supervisor launched (Child PID: {self.supervisor_process.pid})")
 
-    def stop_supervisor(self, timeout: float = 5.0, hard: bool = True) -> None:
-        """Gracefully stop the supervisor process"""
-        if self.supervisor_process and self.supervisor_process.is_alive():
-            self.logger.info("Signaling worker and supervisor to stop...")
+    def stop_worker(self, timeout: Optional[float] = None, hard: bool = False) -> None:
+        if self.worker_process and self.worker_process.is_alive():
+            self.logger.info("Signaling worker to stop...")
             
             # 1. Signal the event
-            self.stop_event.set()
+            self.stop_worker_event.set()
             
+            # 2. Wait for it to finish gracefully
+            self.worker_process.join(timeout=timeout)
+            
+            # 3. Force kill if it's stuck (optional safety net)
+            if hard:
+                self.hard_cleanup(supervisor=False, worker=True)
+            
+            self.logger.info("Worker stopped.")
+        else:
+            self.logger.warning("No worker process to stop")
+            
+            
+    def stop_supervisor(self, timeout: Optional[float] = None, hard: bool = False, and_worker: bool= False) -> None:
+        """Gracefully stop the supervisor process"""
+        if self.supervisor_process and self.supervisor_process.is_alive():
+            self.logger.info("Signaling supervisor to stop...")
+            
+            # 1. Signal the event
+            self.stop_daemon_event.set()
+            
+            if and_worker:
+                self.stop_worker_event.set()
+            
+
             # 2. Wait for it to finish gracefully
             self.supervisor_process.join(timeout=timeout)
             
@@ -161,7 +182,12 @@ class Daemon:
             
             self.local_state.to_disk()
             self.logger.info("Supervisor stopped.")
+
         else:
             self.logger.warning("No supervisor process to stop")
+
+# --- Manager Registration ---
+            
+
